+++ besu-plugins/linea-sequencer/sequencer/src/main/java/net/consensys/linea/sequencer/txselection/LineaTransactionSelectorFactory.java
@@ -16,10 +16,12 @@
 package net.consensys.linea.sequencer.txselection;
 
 import java.util.Map;
+import java.util.Optional;
 
 import net.consensys.linea.config.LineaProfitabilityConfiguration;
 import net.consensys.linea.config.LineaTracerConfiguration;
 import net.consensys.linea.config.LineaTransactionSelectorConfiguration;
+import net.consensys.linea.jsonrpc.JsonRpcManager;
 import net.consensys.linea.plugins.config.LineaL1L2BridgeSharedConfiguration;
 import net.consensys.linea.sequencer.txselection.selectors.LineaTransactionSelector;
 import org.hyperledger.besu.plugin.services.BlockchainService;
@@ -32,6 +34,7 @@ import org.hyperledger.besu.plugin.services.txselection.PluginTransactionSelecto
  */
 public class LineaTransactionSelectorFactory implements PluginTransactionSelectorFactory {
   private final BlockchainService blockchainService;
+  private final Optional<JsonRpcManager> rejectedTxJsonRpcManager;
   private final LineaTransactionSelectorConfiguration txSelectorConfiguration;
   private final LineaL1L2BridgeSharedConfiguration l1L2BridgeConfiguration;
   private final LineaProfitabilityConfiguration profitabilityConfiguration;
@@ -45,13 +48,15 @@ public class LineaTransactionSelectorFactory implements PluginTransactionSelecto
       final LineaL1L2BridgeSharedConfiguration l1L2BridgeConfiguration,
       final LineaProfitabilityConfiguration profitabilityConfiguration,
       final LineaTracerConfiguration tracerConfiguration,
-      final Map<String, Integer> limitsMap) {
+      final Map<String, Integer> limitsMap,
+      final Optional<JsonRpcManager> rejectedTxJsonRpcManager) {
     this.blockchainService = blockchainService;
     this.txSelectorConfiguration = txSelectorConfiguration;
     this.l1L2BridgeConfiguration = l1L2BridgeConfiguration;
     this.profitabilityConfiguration = profitabilityConfiguration;
     this.tracerConfiguration = tracerConfiguration;
     this.limitsMap = limitsMap;
+    this.rejectedTxJsonRpcManager = rejectedTxJsonRpcManager;
   }
 
   @Override
@@ -62,6 +67,7 @@ public class LineaTransactionSelectorFactory implements PluginTransactionSelecto
         l1L2BridgeConfiguration,
         profitabilityConfiguration,
         tracerConfiguration,
-        limitsMap);
+        limitsMap,
+        rejectedTxJsonRpcManager);
   }
 }