+++ besu-plugins/linea-sequencer/sequencer/src/main/java/net/consensys/linea/sequencer/txselection/LineaTransactionSelectorPlugin.java
@@ -22,8 +22,11 @@ import java.util.Optional;
 import com.google.auto.service.AutoService;
 import lombok.extern.slf4j.Slf4j;
 import net.consensys.linea.AbstractLineaRequiredPlugin;
+import net.consensys.linea.config.LineaTransactionSelectorConfiguration;
+import net.consensys.linea.jsonrpc.JsonRpcManager;
 import org.hyperledger.besu.plugin.BesuContext;
 import org.hyperledger.besu.plugin.BesuPlugin;
+import org.hyperledger.besu.plugin.services.BesuConfiguration;
 import org.hyperledger.besu.plugin.services.BlockchainService;
 import org.hyperledger.besu.plugin.services.TransactionSelectionService;
 
@@ -38,6 +41,8 @@ public class LineaTransactionSelectorPlugin extends AbstractLineaRequiredPlugin 
   public static final String NAME = "linea";
   private TransactionSelectionService transactionSelectionService;
   private BlockchainService blockchainService;
+  private Optional<JsonRpcManager> rejectedTxJsonRpcManager = Optional.empty();
+  private BesuConfiguration besuConfiguration;
 
   @Override
   public Optional<String> getName() {
@@ -61,18 +66,38 @@ public class LineaTransactionSelectorPlugin extends AbstractLineaRequiredPlugin 
                 () ->
                     new RuntimeException(
                         "Failed to obtain BlockchainService from the BesuContext."));
+
+    besuConfiguration =
+        context
+            .getService(BesuConfiguration.class)
+            .orElseThrow(
+                () ->
+                    new RuntimeException(
+                        "Failed to obtain BesuConfiguration from the BesuContext."));
   }
 
   @Override
-  public void beforeExternalServices() {
-    super.beforeExternalServices();
+  public void start() {
+    super.start();
+    final LineaTransactionSelectorConfiguration txSelectorConfiguration =
+        transactionSelectorConfiguration();
+    rejectedTxJsonRpcManager =
+        Optional.ofNullable(txSelectorConfiguration.rejectedTxEndpoint())
+            .map(endpoint -> new JsonRpcManager(besuConfiguration.getDataPath(), endpoint).start());
     transactionSelectionService.registerPluginTransactionSelectorFactory(
         new LineaTransactionSelectorFactory(
             blockchainService,
-            transactionSelectorConfiguration(),
+            txSelectorConfiguration,
             l1L2BridgeSharedConfiguration(),
             profitabilityConfiguration(),
             tracerConfiguration(),
-            createLimitModules(tracerConfiguration())));
+            createLimitModules(tracerConfiguration()),
+            rejectedTxJsonRpcManager));
+  }
+
+  @Override
+  public void stop() {
+    super.stop();
+    rejectedTxJsonRpcManager.ifPresent(JsonRpcManager::shutdown);
   }
 }