+++ besu-plugins/linea-sequencer/sequencer/src/test/java/net/consensys/linea/sequencer/txselection/selectors/TestTransactionEvaluationContext.java
@@ -20,7 +20,8 @@ import org.hyperledger.besu.datatypes.Wei;
 import org.hyperledger.besu.plugin.data.ProcessableBlockHeader;
 import org.hyperledger.besu.plugin.services.txselection.TransactionEvaluationContext;
 
-class TestTransactionEvaluationContext implements TransactionEvaluationContext<PendingTransaction> {
+public class TestTransactionEvaluationContext
+    implements TransactionEvaluationContext<PendingTransaction> {
   private ProcessableBlockHeader processableBlockHeader;
   private PendingTransaction pendingTransaction;
   private Wei transactionGasPrice;